chore(src): Reorganize
