+++ src/extensions/Extension.cpp
@@ -1,5 +1,5 @@
 #include "Extension.h"
-#include "../../entrypoint.h"
+#include "../entrypoint.h"
 #include <swiftly-ext/core.h>
 
 typedef void* (*GetExtensionCls)();
@@ -20,19 +20,19 @@ std::string Extension::GetName()
 
 bool Extension::LoadExtension(bool late)
 {
-    #ifdef _WIN32
-        m_hModule = dlmount((std::string("addons/swiftly/extensions/win64/") + m_name + ".dll").c_str());
-    #else
-        m_hModule = dlopen((std::string("addons/swiftly/extensions/linuxsteamrt64/") + m_name + ".so").c_str(), RTLD_NOW);
+#ifdef _WIN32
+    m_hModule = dlmount((std::string("addons/swiftly/extensions/win64/") + m_name + ".dll").c_str());
+#else
+    m_hModule = dlopen((std::string("addons/swiftly/extensions/linuxsteamrt64/") + m_name + ".so").c_str(), RTLD_NOW);
 
-        if(!m_hModule) {
-            PRINTF("Failed to load extension: %s\n", dlerror());
-            return false;
-        }
-    #endif
+    if (!m_hModule) {
+        PRINTF("Failed to load extension: %s\n", dlerror());
+        return false;
+    }
+#endif
 
     void* fnGetClass = reinterpret_cast<void*>(dlsym(m_hModule, "GetExtensionClass"));
-    if(!fnGetClass) {
+    if (!fnGetClass) {
         PRINTF("Failed to load extension: GetExtensionClass is not present.\n", dlerror());
         return false;
     }
@@ -42,7 +42,7 @@ bool Extension::LoadExtension(bool late)
     std::string err;
     bool res = m_api->Load(err, g_SHPtr, g_SMAPI, late);
     loaded = res;
-    if(!res) {
+    if (!res) {
         PRINT("An error has occured while trying to load the extension.\n");
         PRINTF("Extension: %s.\n", m_name.c_str());
         PRINTF("Error: %s\n", err.c_str());
@@ -58,7 +58,7 @@ bool Extension::UnloadExtension()
 {
     std::string err;
     bool res = m_api->Unload(err);
-    if(!res) {
+    if (!res) {
         PRINT("An error has occured while trying to unload the extension.\n");
         PRINTF("Extension: %s.\n", m_name.c_str());
         PRINTF("Error: %s\n", err.c_str());
