+++ src/extensions/ExtensionManager.h
